Refactor request resource tests.
